mcp-server/requirements.txt
# MCP Server for Logan Security Dashboard
# Core MCP dependencies
mcp>=1.0.0

# OCI SDK
oci>=2.126.0

# Standard library extensions
typing-extensions>=4.0.0
pydantic>=2.0.0

# Async support
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Logging and utilities
python-dateutil>=2.8.0
pytz>=2023.3

# JSON handling
orjson>=3.9.0

# HTTP client for external APIs
httpx>=0.25.0
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Security
cryptography>=41.0.0

# Configuration
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0